Added renderer concept
